pyproject.toml
[project]
name = "python-base-project"
version = "0.1.0"
description = "Type here the description for your project."
readme = "README.md"
requires-python = "==3.11.9"
dependencies = [
    "aws-cdk-lib>=2.0.0",
    "constructs>=10.0.0",
    "pre-commit>=4.0.1",
    "pytest>=8.3.3",
    "python-dotenv==1.0.1",
    "setuptools>=75.6.0",
    "wheel>=0.45.1",
]

[tool.uv]
dev-dependencies = [
    "black==24.10.0",
    "pyright>=1.1.388",
    "pytest==8.3.3",
    "ruff==0.7.1",
    "typos==1.26.8",
    "pre-commit"
]



[tool.ruff]
src = ["app"]
line-length = 120
show-fixes = true
unsafe-fixes = true

[tool.ruff.lint]
select = ["ALL"]
ignore = ["COM812", "ISC001", "TCH001", "G002"]

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint.pycodestyle]
max-doc-length = 120

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.pylint]
max-args = 10

[tool.pyright]
include = ["app"]  # Adjust this to your source directory
exclude = [
    "**/.venv",
    "**/__pycache__",
    "**/tests"
]
typeCheckingMode = "strict"  # For strictest type checking
reportMissingTypeStubs = true
reportUnknownParameterType = true
reportUnknownArgumentType = true
reportUnknownLambdaType = true
reportUnknownVariableType = true
reportUnknownMemberType = true
reportMissingTypeArgument = true
reportInvalidTypeVarUse = true
reportUnnecessaryCast = true
reportUnnecessaryComparison = true
reportUnnecessaryContains = true
reportCallInDefaultInitializer = true
reportPropertyTypeMismatch = true
reportWildcardImportFromLibrary = true
reportUntypedFunctionDecorator = true
reportUntypedClassDecorator = true
reportUntypedBaseClass = true
useLibraryCodeForTypes = true
venvPath = "."
venv = ".venv"

[tool.black]
line-length = 120
target-version = ['py311']

[tool.ruff.lint.isort]
combine-as-imports = true
split-on-trailing-comma = true
relative-imports-order = "closest-to-furthest"
force-single-line = false
force-wrap-aliases = false
